quantum add is now working properly

+++ include/Matrix.h
@@ -9,7 +9,7 @@
 #include <vector>
 using std::vector;
 
-typedef std::complex<double> Complex;
+typedef std::complex<float> Complex;
 
 // квадратная матрица
 class Matrix {
